Removed old files and added missed media query
